test: consolidate sequence number tests into their natural homes
Move metadata persistence tests to AnalyticsExporterTest and OTel
pipeline tests to OtelSdkManagerTest. AnalyticsExporterSequenceNumberTest
was a temporary staging area that no longer had a clear responsibility.
